+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * upd charms: add position parameter to show charm functions
 * upd charms: add function to check charm is open
 * additional colors moved to metro-colors.css
+* add validator multiple funcs, see validator multi func demo
 
 ## 3.0.14
 * fix tabcontrol frames overflow